refactor: restructure recurrence functionality and improve code organization
