+++ apps/mobile/app/recorrencia/[id].style.ts
@@ -1,0 +1,35 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  header: {
+    alignSelf: 'stretch',
+    width: '100%',
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingHorizontal: 16,
+    paddingTop: 56,
+    paddingBottom: 14,
+    borderBottomWidth: 1,
+  },
+  headerBtn: {
+    width: 40,
+    height: 40,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  headerTitle: {
+    fontSize: 18,
+    fontWeight: '700',
+    flex: 1,
+    textAlign: 'center',
+  },
+  content: {
+    paddingHorizontal: 16,
+    paddingTop: 16,
+    paddingBottom: 80,
+    gap: 16,
+    alignItems: 'stretch',
+    width: '100%',
+  },
+})
