+++ apps/mobile/app/recorrencia/index.style.ts
@@ -1,0 +1,50 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  scroll: {
+    alignSelf: 'stretch',
+    width: '100%',
+  },
+  content: {
+    paddingHorizontal: 8,
+    paddingTop: 48,
+    paddingBottom: 100,
+    gap: 14,
+    alignItems: 'stretch',
+    width: '100%',
+  },
+  header: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    alignSelf: 'stretch',
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: '700',
+    flex: 1,
+    textAlign: 'center',
+  },
+  iconBtn: {
+    width: 38,
+    height: 38,
+    borderRadius: 19,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  fab: {
+    position: 'absolute',
+    bottom: 28,
+    right: 24,
+    width: 56,
+    height: 56,
+    borderRadius: 28,
+    alignItems: 'center',
+    justifyContent: 'center',
+    elevation: 6,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 3 },
+    shadowOpacity: 0.25,
+    shadowRadius: 6,
+  },
+})
