+++ apps/mobile/app/recorrencia/nova.style.ts
@@ -1,0 +1,137 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  screen: {
+    flex: 1,
+    alignSelf: 'stretch',
+    width: '100%',
+  },
+  header: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    paddingHorizontal: 20,
+    paddingTop: 52,
+    paddingBottom: 12,
+  },
+  headerBtn: {
+    width: 40,
+    height: 40,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  headerTitle: {
+    fontSize: 17,
+    fontWeight: '600',
+    flex: 1,
+    textAlign: 'center',
+  },
+  content: {
+    paddingHorizontal: 20,
+    paddingTop: 4,
+    gap: 16,
+  },
+  titleBlock: {
+    gap: 4,
+    paddingHorizontal: 20,
+    paddingBottom: 8,
+  },
+  title: {
+    fontSize: 26,
+    fontWeight: '700',
+  },
+  subtitle: {
+    fontSize: 13,
+  },
+  typeTabs: {
+    flexDirection: 'row',
+    borderRadius: 10,
+    padding: 4,
+    marginHorizontal: 20,
+  },
+  typeTab: {
+    flex: 1,
+    paddingVertical: 10,
+    borderRadius: 8,
+    alignItems: 'center',
+  },
+  typeTabText: {
+    fontSize: 13,
+    fontWeight: '700',
+    letterSpacing: 0.5,
+  },
+  fieldGroup: {
+    gap: 6,
+  },
+  label: {
+    fontSize: 11,
+    fontWeight: '700',
+    letterSpacing: 0.8,
+  },
+  fieldInput: {
+    borderRadius: 12,
+    paddingHorizontal: 14,
+    paddingVertical: 14,
+    fontSize: 15,
+  },
+  fieldError: {
+    borderWidth: 1,
+  },
+  dropdownRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderRadius: 12,
+    paddingHorizontal: 14,
+    paddingVertical: 14,
+    gap: 8,
+  },
+  dropdownText: {
+    fontSize: 15,
+  },
+  pickerList: {
+    borderRadius: 12,
+    overflow: 'hidden',
+  },
+  pickerItem: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 12,
+    paddingHorizontal: 16,
+    paddingVertical: 13,
+  },
+  pickerItemText: {
+    fontSize: 15,
+    fontWeight: '500',
+  },
+  toggleRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    borderRadius: 12,
+    paddingHorizontal: 16,
+    paddingVertical: 16,
+  },
+  toggleInfo: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 12,
+    flex: 1,
+  },
+  toggleTextBlock: {
+    flex: 1,
+    gap: 3,
+  },
+  toggleTitle: {
+    fontSize: 14,
+    fontWeight: '600',
+  },
+  toggleValue: {
+    fontSize: 13,
+  },
+  toggleSub: {
+    fontSize: 12,
+  },
+  bottomSpacer: {
+    height: 32,
+  },
+})
