+++ apps/mobile/src/components/finance/recurrences/RecurrenceDetailsCard/style.ts
@@ -1,0 +1,104 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  mainCard: {
+    borderRadius: 20,
+    paddingHorizontal: 20,
+    paddingVertical: 20,
+    gap: 16,
+    alignSelf: 'stretch',
+  },
+  cardTop: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 14,
+  },
+  iconCircle: {
+    width: 56,
+    height: 56,
+    borderRadius: 28,
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexShrink: 0,
+  },
+  cardTopText: {
+    flex: 1,
+    gap: 4,
+  },
+  typeLabel: {
+    fontSize: 11,
+    fontWeight: '700',
+    letterSpacing: 0.8,
+  },
+  descriptionText: {
+    fontSize: 18,
+    fontWeight: '700',
+    letterSpacing: -0.2,
+  },
+  divider: {
+    height: 1,
+    marginHorizontal: -20,
+  },
+  amountRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    gap: 12,
+  },
+  amountBlock: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'flex-end',
+    gap: 4,
+  },
+  amountCurrency: {
+    fontSize: 18,
+    fontWeight: '500',
+    paddingBottom: 4,
+  },
+  amountNumber: {
+    fontSize: 36,
+    fontWeight: '700',
+    letterSpacing: -1,
+    lineHeight: 40,
+  },
+  statusBadge: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 6,
+    paddingHorizontal: 12,
+    paddingVertical: 6,
+    borderRadius: 20,
+    flexShrink: 0,
+  },
+  statusDot: {
+    width: 7,
+    height: 7,
+    borderRadius: 4,
+  },
+  statusText: {
+    fontSize: 11,
+    fontWeight: '700',
+    letterSpacing: 0.5,
+  },
+  infoBlock: {
+    gap: 12,
+  },
+  btnRow: {
+    flexDirection: 'row',
+    gap: 12,
+  },
+  btn: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: 8,
+    paddingVertical: 14,
+    borderRadius: 12,
+  },
+  btnText: {
+    fontSize: 15,
+    fontWeight: '700',
+  },
+})
