--- apps/mobile/src/components/finance/recurrences/recurrences-store.ts
@@ -1,11 +1,0 @@
-let _pendingDeleteId: string | null = null
-
-export function setPendingDeleteId(id: string | null) {
-  _pendingDeleteId = id
-}
-
-export function consumePendingDeleteId(): string | null {
-  const id = _pendingDeleteId
-  _pendingDeleteId = null
-  return id
-}
